Refactor mutation to be separate functions
